LayoutTests/imported/w3c/web-platform-tests/css/css-position/position-relative-aspect-ratio-002-expected.txt


PASS Document is in quirks mode
PASS Quirks: explicit width + aspect-ratio 2/1, top: 50%
PASS Quirks: auto width block + aspect-ratio 2/1, top: 50%
